backend/requirements-dev.txt
-r requirements.txt
black>=23.0.0
flake8>=6.0.0
mypy>=1.5.0
pytest-mock>=3.11.0
